cJSON uses a range hack to decide whether a number should be treated internally as a double or an int. The 64-bit int solution I implemented replaced the internal 32-bit int with a 64-bit int.
There is a weakness to the cJSON design that if the code opts for using
double, it doesn’t remember how many decimal places the original JSON
had and may have precision issues. This is visible in the test harness
as FP lat/long values are output with many more decimal places of
precision than the JSON input had. This issue doesn’t happen with
64-bit, just FP.

I added a 64-bit test to the end of test.c::

int main()
{	...
        char text6[]="{ \"big int\":  4000000000, \"minus10\": -10000000000  }";
        puts("long long int test:");
        doit(text6);
        return 0;
}

./test OUTPUT:

long long int test:
{
	"big int":	4000000000,
	"minus10":	-10000000000
}

All I tested was this simple test.  There’s also a ‘tests’ directory
that seems to be a duplicate of the main directory. I didn’t do
anything with that.